refactor: 식당 application 패키지 분리
