+++ src/main/java/com/michelet/restaurantservice/restaurant/application/result/RestaurantSummaryResult.java
@@ -1,7 +1,7 @@
-package com.michelet.restaurant.application.result;
+package com.michelet.restaurantservice.restaurant.application.result;
 
-import com.michelet.restaurant.domain.model.Restaurant;
-import com.michelet.restaurant.domain.model.RestaurantStatus;
+import com.michelet.restaurantservice.restaurant.domain.model.Restaurant;
+import com.michelet.restaurantservice.restaurant.domain.model.RestaurantStatus;
 
 import java.util.UUID;
